Fetch server paths from Nginx configuration files.
This patch changes the way the `$path` is fetched for each `$host`.

They are both completely different entities, `$path` does not necessarily contain `$host`.

> Example:
> `$host` might be `localsite.dev` whereas `$path` could be `finalserver.com`

The `$path` is fetched from the Nginx configuration files, which means that there's no need to have any special cases hard-coded into the source, like the default sites that come with VVV.